+++ blueprints/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/formula_8_27.py
@@ -1,6 +1,6 @@
-"""Formula 8.27 from EN 1993-1-1:2025: Chapter 8 - Ultimate Limit State."""
+"""Formula 8.27 from EN 1993-1-1:2022: Chapter 8 - Ultimate Limit State."""
 
-from blueprints.codes.eurocode.en_1993_1_1_2025 import EN_1993_1_1_2025
+from blueprints.codes.eurocode.en_1993_1_1_2022 import EN_1993_1_1_2022
 from blueprints.codes.formula import Formula
 from blueprints.codes.latex_formula import LatexFormula, latex_replace_symbols
 from blueprints.type_alias import DIMENSIONLESS, MM
@@ -11,7 +11,7 @@ class Form8Dot27CheckShearBucklingResistance(Formula):
     r"""Class representing formula 8.27 for checking shear buckling resistance for webs without intermediate stiffeners."""
 
     label = "8.27"
-    source_document = EN_1993_1_1_2025
+    source_document = EN_1993_1_1_2022
 
     def __init__(
         self,
@@ -22,7 +22,7 @@ class Form8Dot27CheckShearBucklingResistance(Formula):
     ) -> None:
         r"""Check the shear buckling resistance for webs without intermediate stiffeners.
 
-        EN 1993-1-1:2025 art.8.2.6(6) - Formula (8.27)
+        EN 1993-1-1:2022 art.8.2.6(6) - Formula (8.27)
 
         Parameters
         ----------
