+++ tests/codes/eurocode/en_1993_1_1_2022/chapter_8_ultimate_limit_state/test_formula_8_28.py
@@ -1,13 +1,13 @@
-"""Testing formula 8.28 from EN 1993-1-1:2025, chapter 8, ultimate limit state."""
+"""Testing formula 8.28 from EN 1993-1-1:2022, chapter 8, ultimate limit state."""
 
 import pytest
 
-from blueprints.codes.eurocode.en_1993_1_1_2025.chapter_8_ultimate_limit_state.formula_8_28 import Form8Dot28TotalTorsionalMoment
+from blueprints.codes.eurocode.en_1993_1_1_2022.chapter_8_ultimate_limit_state.formula_8_28 import Form8Dot28TotalTorsionalMoment
 from blueprints.validations import NegativeValueError
 
 
 class TestForm8Dot28TotalTorsionalMoment:
-    """Validation for formula 8.28 from EN 1993-1-1:2025, chapter 8, ultimate limit state."""
+    """Validation for formula 8.28 from EN 1993-1-1:2022, chapter 8, ultimate limit state."""
 
     def test_evaluation(self) -> None:
         """Tests the evaluation of the result."""
